# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TravelMind API",
    "description": "TravelMind dynamic travel planning agent API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/trips": {
      "post": {
        "tags": [
          "trips"
        ],
        "summary": "Create Trip Endpoint",
        "description": "创建新旅行的API端点",
        "operationId": "created_trip",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trips/{trip_id}": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "Get Trip Response",
        "operationId": "get_trip",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trips/{trip_id}/planning-runs": {
      "post": {
        "tags": [
          "planning"
        ],
        "summary": "Start Planning Endpoint",
        "operationId": "start_planning_run",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartPlanningResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trips/{trip_id}/planning-runs/{run_id}/events": {
      "get": {
        "tags": [
          "planning"
        ],
        "summary": "List Planning Events",
        "operationId": "list_planning_events",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanningEventListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trips/{trip_id}/plans": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "List Plan Versions",
        "operationId": "list_plan_versions",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trips/{trip_id}/plans/{version}": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "Get Plan Endpoint",
        "operationId": "get_plan_version",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trips/{trip_id}/feedback": {
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Create Feedback Endpoint",
        "description": "保存结构化反馈，并按需启动下一版规划。\n\n如果反馈缺少明确操作，则返回 200 并要求用户澄清；\n否则正常返回 202 和可选的重新规划任务。",
        "operationId": "create_feedback",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trips/{trip_id}/manual-edits": {
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "Create Manual Plan Edit",
        "operationId": "create_manual_plan_edit",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualPlanEditRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualPlanEditResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CurrentPlanResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "trip_id": {
            "type": "string",
            "format": "uuid",
            "title": "Trip Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "parent_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Version"
          },
          "status": {
            "$ref": "#/components/schemas/PlanStatus"
          },
          "itinerary": {
            "$ref": "#/components/schemas/Itinerary"
          },
          "constraint_report": {
            "$ref": "#/components/schemas/ConstraintReport"
          },
          "change_summary": {
            "type": "string",
            "title": "Change Summary"
          },
          "trigger": {
            "$ref": "#/components/schemas/PlanTrigger"
          },
          "planning_run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Planning Run Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "accepted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepted At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "trip_id",
          "version",
          "parent_version",
          "status",
          "itinerary",
          "constraint_report",
          "change_summary",
          "trigger",
          "planning_run_id",
          "created_at",
          "accepted_at"
        ],
        "title": "CurrentPlanResponse"
      },
...
      "DateRange": {
        "properties": {
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "start_date",
          "end_date"
        ],
        "title": "DateRange",
        "description": "日期区间\n用于旅行的行程起止时间"
      },
...
      "FeedbackCreateRequest": {
        "properties": {
          "base_plan_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Base Plan Version"
          },
          "message": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message"
          },
          "client_operations": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/SetBudgetOperation"
                },
                {
                  "$ref": "#/components/schemas/SetMaxWalkingOperation"
                },
                {
                  "$ref": "#/components/schemas/AddRequiredPlaceOperation"
                },
                {
                  "$ref": "#/components/schemas/AddExcludedPlaceOperation"
                }
              ],
              "discriminator": {
                "propertyName": "op",
                "mapping": {
                  "add_excluded_place": "#/components/schemas/AddExcludedPlaceOperation",
                  "add_required_place": "#/components/schemas/AddRequiredPlaceOperation",
                  "set_budget": "#/components/schemas/SetBudgetOperation",
                  "set_max_walking": "#/components/schemas/SetMaxWalkingOperation"
                }
              }
            },
            "type": "array",
            "title": "Client Operations"
          },
          "auto_start_replanning": {
            "type": "boolean",
            "title": "Auto Start Replanning",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "base_plan_version",
          "message"
        ],
        "title": "FeedbackCreateRequest"
      },
...
      "FeedbackResponse": {
        "properties": {
          "feedback": {
            "$ref": "#/components/schemas/FeedbackRecordResponse"
          },
          "planning_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlanningRunResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "events_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Events Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "feedback",
          "planning_run",
          "events_url"
        ],
        "title": "FeedbackResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManualPlanEditRequest": {
        "properties": {
          "base_plan_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Base Plan Version"
          },
          "days": {
            "items": {
              "$ref": "#/components/schemas/ManualDayEdit"
            },
            "type": "array",
            "minItems": 1,
            "title": "Days"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "base_plan_version",
          "days"
        ],
        "title": "ManualPlanEditRequest"
      },
      "ManualPlanEditResponse": {
        "properties": {
          "plan": {
            "$ref": "#/components/schemas/CurrentPlanResponse"
          },
          "planning_run": {
            "$ref": "#/components/schemas/PlanningRunResponse"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "plan",
          "planning_run"
        ],
        "title": "ManualPlanEditResponse"
      },
...
      "PlanListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlanVersionSummary"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "PlanListResponse"
      },
      "PlanStatus": {
        "type": "string",
        "enum": [
          "draft",
          "valid",
          "accepted",
          "superseded"
        ],
        "title": "PlanStatus"
      },
...
      "PlanningEventListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlanningEventResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "PlanningEventListResponse"
      },
...
      "StartPlanningResponse": {
        "properties": {
          "planning_run": {
            "$ref": "#/components/schemas/PlanningRunResponse"
          },
          "events_url": {
            "type": "string",
            "title": "Events Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "planning_run",
          "events_url"
        ],
        "title": "StartPlanningResponse"
      },
...
      "TripConstraints": {
        "properties": {
          "total_budget": {
            "$ref": "#/components/schemas/Money"
          },
          "budget_is_hard_limit": {
            "type": "boolean",
            "title": "Budget Is Hard Limit",
            "default": true
          },
          "daily_start_time": {
            "type": "string",
            "pattern": "^(?:[01]\\d|2[0-3]):[0-5]\\d$",
            "title": "Daily Start Time",
            "default": "09:00"
          },
          "daily_end_time": {
            "type": "string",
            "pattern": "^(?:[01]\\d|2[0-3]):[0-5]\\d$",
            "title": "Daily End Time",
            "default": "21:00"
          },
          "max_walking_meters_per_day": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 50000.0,
                "minimum": 1000.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Walking Meters Per Day",
            "default": 12000
          },
          "max_activities_per_day": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Max Activities Per Day",
            "default": 5
          },
          "minimum_transfer_buffer_minutes": {
            "type": "integer",
            "maximum": 60.0,
            "minimum": 0.0,
            "title": "Minimum Transfer Buffer Minutes",
            "default": 10
          },
          "rest_minutes_per_day": {
            "type": "integer",
            "maximum": 240.0,
            "minimum": 0.0,
            "title": "Rest Minutes Per Day",
            "default": 60
          },
          "required_place_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Place Names"
          },
          "excluded_place_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Excluded Place Names"
          },
          "accessible_only": {
            "type": "boolean",
            "title": "Accessible Only",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "total_budget"
        ],
        "title": "TripConstraints",
        "description": "行程硬性约束条件：算法必须遵守，不满足就不能生成方案"
      },
      "TripCreateRequest": {
        "properties": {
          "origin": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Origin"
          },
          "destination": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Destination"
          },
          "destination_timezone": {
            "type": "string",
            "title": "Destination Timezone"
          },
          "date_range": {
            "$ref": "#/components/schemas/DateRange"
          },
          "travelers": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 1.0,
            "title": "Travelers"
          },
          "preferences": {
            "$ref": "#/components/schemas/TripPreferences"
          },
          "constraints": {
            "$ref": "#/components/schemas/TripConstraints"
          },
          "locale": {
            "type": "string",
            "pattern": "^[a-z]{2,3}(?:-[A-Z]{2})?$",
            "title": "Locale",
            "default": "zh-CN"
          },
          "display_currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$",
            "title": "Display Currency",
            "default": "CNY"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "origin",
          "destination",
          "destination_timezone",
          "date_range",
          "travelers",
          "preferences",
          "constraints"
        ],
        "title": "TripCreateRequest",
        "description": "创建旅行请求。\n\n它和领域 TripRequest 当前字段相同，但拥有独立的 OpenAPI Schema 名称。\n客户端无法提交 id、status、revision 或 current_plan_version。"
      },
      "TripPreferences": {
        "properties": {
          "interests": {
            "items": {
              "$ref": "#/components/schemas/WeightedPreference"
            },
            "type": "array",
            "title": "Interests"
          },
          "avoid": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Avoid"
          },
          "dietary": {
            "items": {
              "$ref": "#/components/schemas/DietaryPreference"
            },
            "type": "array",
            "title": "Dietary"
          },
          "transport_modes": {
            "items": {
              "$ref": "#/components/schemas/TransportMode"
            },
            "type": "array",
            "title": "Transport Modes"
          },
          "accommodation_notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Accommodation Notes"
          },
          "pace": {
            "$ref": "#/components/schemas/Pace",
            "default": "balanced"
          },
          "must_visit_place_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Must Visit Place Names"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TripPreferences",
        "description": "用户行程偏好：主观喜好，不做硬性强制约束，供算法做加权推荐"
      },
      "TripResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/TripStatus"
          },
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision"
          },
          "origin": {
            "type": "string",
            "title": "Origin"
          },
          "destination": {
            "type": "string",
            "title": "Destination"
          },
          "destination_timezone": {
            "type": "string",
            "title": "Destination Timezone"
          },
          "date_range": {
            "$ref": "#/components/schemas/DateRange"
          },
          "travelers": {
            "type": "integer",
            "title": "Travelers"
          },
          "preferences": {
            "$ref": "#/components/schemas/TripPreferences"
          },
          "constraints": {
            "$ref": "#/components/schemas/TripConstraints"
          },
          "locale": {
            "type": "string",
            "title": "Locale"
          },
          "display_currency": {
            "type": "string",
            "title": "Display Currency"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "current_plan_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Plan Version"
          },
          "active_planning_run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Planning Run Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "status",
          "revision",
          "origin",
          "destination",
          "destination_timezone",
          "date_range",
          "travelers",
          "preferences",
          "constraints",
          "locale",
          "display_currency",
          "notes",
          "current_plan_version",
          "active_planning_run_id",
          "created_at",
          "updated_at"
        ],
        "title": "TripResponse"
      },
      "TripStatus": {
        "type": "string",
        "enum": [
          "draft",
          "planning",
          "needs_review",
          "replanning",
          "completed",
          "failed",
          "archived"
        ],
        "title": "TripStatus"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}